UHF-9832: Combined user sanitize form and user sanitize drush command tests.
